Added ConfigureGenerator component
